backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
reportlab>=4.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic[email]>=2.0.0
pytesseract>=0.3.10
pillow>=10.0.0
openai>=1.0.0
python-pptx>=0.7.11
comtypes>=1.1.14
python-dotenv>=1.0.0
gspread>=5.10.0
oauth2client>=4.1.3
